# Unicode-creation optimizations