--- a/microbenchmarks/re_finditer_bench.py
+++ b/microbenchmarks/re_finditer_bench.py
@@ -1,0 +1,7 @@
+import re
+def f():
+    r = re.compile(" ")
+    u = "a b c d"
+    for i in xrange(2000000):
+        r.finditer(u)
+f()