--- a/microbenchmarks/unicode_split_ubench.py
+++ b/microbenchmarks/unicode_split_ubench.py
@@ -1,0 +1,6 @@
+def f():
+    u = "a b c d"
+    u2 = u" "
+    for i in xrange(4000000):
+        u.split(u2)
+f()