Mostly, inline things into _PyUnicode_New to reduce its overhead.

Also, sneak in a change to reenable PyString_GET_SIZE.

Overall effect:
```
       django_template.py             4.7s (2)             4.6s (2)  -2.1%
            pyxl_bench.py             4.0s (2)             4.0s (2)  -0.4%
sqlalchemy_imperative2.py             5.3s (2)             5.3s (2)  -1.1%
        django_migrate.py             2.0s (2)             2.0s (2)  -2.0%
      virtualenv_bench.py             5.3s (2)             5.3s (2)  -0.4%
                  geomean                 4.0s                 4.0s  -1.2%
```